modify setup

+++ setup.sh
@@ -1,5 +1,12 @@
 #!/bin/bash
-echo "开始配置和安装vim......"
+echo "开始安装和配置vim......"
 
-mv -f ~/.vimrc ~/.vimrc.back
+if [ -f "~/.vimrc" ]; then
+	mv -f ~/.vimrc ~/.vimrc.back
+fi
 cp -f .vimrc ~/.vimrc
+
+vimDir="/home/wenba/.vim/"
+if [ ! -d "$vimDir" ]; then
+	mkdir "$vimDir"
+fi
